ConnectionService AddResource update

+++ src/apps/Altinn.AccessManagement/src/Altinn.AccessManagement.Core/Models/SimpleDelegationChange.cs
@@ -11,7 +11,7 @@ namespace Altinn.AccessManagement.Core.Models
         /// Gets or sets the delegation change id
         /// </summary>
         [JsonPropertyName("i")]
-        public long DelegationChangeId { get; set; }
+        public int DelegationChangeId { get; set; }
 
         /// <summary>
         /// Gets or sets the altinnappid. E.g. skd/skattemelding
